+++ src/hmmat.h
@@ -25,7 +25,7 @@ public:
   HmMat() : m_dendX(NULL), m_dendY(NULL) {}
   ~HmMat();
   bool read(const char* fn);
-  float get(uint32_t row, uint32_t col) const { return m_d[row][col]; }
+  double get(uint32_t row, uint32_t col) const { return m_d[row][col]; }
   char   isSel(uint32_t row, uint32_t col) const { return m_sel[row][col]; }
   void   sel(uint32_t row, uint32_t col) { m_sel[row][col] = 1; }
   void   sel(uint32_t r0, uint32_t c0, uint32_t r1, uint32_t c1);
@@ -36,8 +36,8 @@ public:
   uint32_t nrow() const { return m_d.size(); }
   uint32_t ncol() const { return m_d.empty() ? 0 : m_d.front().size(); }
   void print() const;
-  float maxVal() const;
-  float minVal() const;
+  double maxVal() const;
+  double minVal() const;
   void transpose();
   void order(OrderType ot);
   const Dendrogram *getDendY() const { return m_dendY; }
@@ -52,7 +52,7 @@ private:
   void orderRandomY();
   void resetOrderX();
   void resetOrderY();
-  std::vector<std::vector<float>> m_d;
+  std::vector<std::vector<double>> m_d;
   std::vector<std::vector<char>> m_sel;
   std::vector<std::string> m_xlab;
   std::vector<std::string> m_ylab;
